refactor: tservers are utils
