+++ src/tests/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './tclient';
+export * from './tservers';
